refactor: organize server modules
